fix: babysit round 1, 15 findings (13 coderabbit, 2 gemini)
Gemini (2): reorder raw_cached check before app_state.slice() in _service
and _delete_entry_service so SimpleNamespace test app-states without a
slice method are served via the documented fallback instead of crashing.

CodeRabbit (13):
- shutdown_checkpoint: Final[float] on timeout constants.
- _proposal_models / _result_models: add extra=forbid to frozen ConfigDict.
- extract _mcp_handler_common.py for the 4 helpers duplicated across the
  organization + integrations handler helper modules; update consumers.
- agents_crud: drop redundant function-local AgentIdentity import.
- add reraise_critical(exc) to generic except blocks across
  integrations_artifacts_ontology, integrations_catalog,
  integrations_oauth_clients, organization_company, organization_departments,
  organization_teams_roles, and the one missing memory_finetune handler.
- code_applier: bind OSError as exc and log error_type + safe_error_description.
- architecture_applier: extract dry_run per-change validation into a helper.
