+++ src/synthorg/meta/mcp/handlers/integrations_catalog.py
@@ -9,12 +9,13 @@ emitting ``MCP_ADMIN_OP_EXECUTED`` on success.
 from typing import TYPE_CHECKING, Any
 
 from synthorg.communication.mcp_errors import CapabilityNotSupportedError
+from synthorg.core.critical_errors import reraise_critical
 from synthorg.infrastructure.state import mcp_catalog_facade_service_of
 from synthorg.meta.mcp.errors import (
     ArgumentValidationError,
     GuardrailViolationError,
 )
-from synthorg.meta.mcp.handlers._integrations_helpers import (
+from synthorg.meta.mcp.handlers._mcp_handler_common import (
     _map_capability,
     _require_str,
     _to_jsonable,
@@ -64,6 +65,7 @@ async def _mcp_catalog_list(
         log_handler_argument_invalid(tool, exc)
         return err(exc)
     except Exception as exc:
+        reraise_critical(exc)
         log_handler_invoke_failed(tool, exc)
         return err(exc)
     sequence = tuple(entries)
@@ -97,6 +99,7 @@ async def _mcp_catalog_search(
         log_handler_argument_invalid(tool, exc)
         return err(exc)
     except Exception as exc:
+        reraise_critical(exc)
         log_handler_invoke_failed(tool, exc)
         return err(exc)
     return ok([_to_jsonable(e) for e in entries])
@@ -125,6 +128,7 @@ async def _mcp_catalog_get(
         log_handler_argument_invalid(tool, exc)
         return err(exc)
     except Exception as exc:
+        reraise_critical(exc)
         log_handler_invoke_failed(tool, exc)
         return err(exc)
     if entry is None:
@@ -160,6 +164,7 @@ async def _mcp_catalog_install(
         log_handler_argument_invalid(tool, exc)
         return err(exc)
     except Exception as exc:
+        reraise_critical(exc)
         log_handler_invoke_failed(tool, exc)
         return err(exc)
     return ok(_to_jsonable(result))
@@ -205,6 +210,7 @@ async def _mcp_catalog_uninstall(
         log_handler_argument_invalid(tool, exc)
         return err(exc)
     except Exception as exc:
+        reraise_critical(exc)
         log_handler_invoke_failed(tool, exc)
         return err(exc)
     return ok({"removed": removed})
